rename : 🚚 GlobalStyle 파일명 변경
