+++ week3/assignment/점메추 🍚/src/components/MainCharacter.jsx
@@ -1,5 +1,5 @@
-import { Character } from "../../styles/globalStyle";
-import mainCharacter from "../../assets/img/character.png";
+import { Character } from "../styles/GlobalStyle";
+import mainCharacter from "../assets/img/character.png";
 
 export default function MainCharacter() {
   return <Character src={mainCharacter} alt="캐릭터" />;
