+++ week3/assignment/점메추 🍚/src/main.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
-import GlobalStyle from "./styles/globalStyle.js";
+import GlobalStyle from "./styles/GlobalStyle.js";
 import { ThemeProvider } from "styled-components";
 import theme from "./styles/theme.js";
